Improve UI
I found it kinda annoying when I opened the task editor today
and the design choices left it feeling very amateurish.

The biggest issue I think was the lack of any indication of grouping. Even
though the columns were fairly easy to spot, it was still easy to lose
track of them.
The next most annoying thing I noticed was the lack of separation of the
control of what view you were looking at and the view itself.

The third was that the category text edit field was not very clear, even
with the hover text, being separated by a huge swath of space.

I don't like that you can't select the current value of the category in the combobox,
but making it possible to do that required an allocation each frame for what
felt like an exceptionally minimal improvement at the moment(I don't feel like
giving it a thread local static storage this time)

(I'm trying out the 'keep the commit messages like a lab journal' advice
I saw on the fediverse)