Save files on a separate thread
This is most likely not that big an improvement, but it felt like
one of those best practices that I should try to follow. After all,
just because I have an SSD doesn't mean that someone isn't running this on
a 1000rpm harddrive.

In the future I believe it will be beneficial to pass the results of saving
the file. Error or OK should be enough.